+++ src/db/migrations/0007_shop_total.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `shops` ADD `total_minor` integer;