+++ src/db/migrations/migrations.js
@@ -8,6 +8,7 @@ import m0003 from './0003_item_note_urgent.sql';
 import m0004 from './0004_item_not_found.sql';
 import m0005 from './0005_list_look.sql';
 import m0006 from './0006_item_price.sql';
+import m0007 from './0007_shop_total.sql';
 
   export default {
     journal,
@@ -18,7 +19,8 @@ m0002,
 m0003,
 m0004,
 m0005,
-m0006
+m0006,
+m0007
     }
   }